Update Collections.js
It still needs an HTML / CSS tweaking in order to display NFTs as we want them to be presented: NFT/COLLECTION/SIDE/TIER/SERIES/EDITION/POWER (which corresponds to an infrastructural design, abiding to the WhitePaper, Strategic Planning, and Branding), so the UX is always at its best, within our current possibilities, meaning that navigation over our website should always be intuitive, easy, fun and entertaining. Displaying a whole bunch of images (which are also distorted by some styling we need to correct). Overall I think it's looking fine.